chore(dispatcher): moves the console dispatcher to its own class (#71).
