+++ bundlers/haystack-agent/src/main/resources/META-INF/services/com.expedia.www.haystack.agent.core.Dispatcher
@@ -1,2 +1,3 @@
 com.expedia.www.haystack.agent.dispatcher.KafkaDispatcher
 com.expedia.www.haystack.agent.dispatcher.KinesisDispatcher
+com.expedia.www.haystack.agent.core.dispatchers.ConsoleDispatcher
